+++ src/test/java/christmas/domain/StartDateTest.java
@@ -1,6 +1,5 @@
 package christmas.domain;
 
-import christmas.domain.contants.StarDate;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
